agent_app/pyproject.toml
[project]
name = "agent-app"
version = "0.1.0"

dependencies = [
    "huggingface_hub",
    "python-dotenv",
]
